Add ginkgo.skip option, reduce code repetition, set Invalid state during Gateway name conflict, general code cleanliness improvements
