--- pkg/model/core/status_condition.go
@@ -1,5 +1,0 @@
-package core
-
-import gwv1 "sigs.k8s.io/gateway-api/apis/v1"
-
-const GatewayReasonConflicted gwv1.GatewayConditionReason = "Conflicted"
