+++ test/suites/integration/suite_test.go
@@ -2,6 +2,9 @@ package integration
 
 import (
 	"context"
+	anaws "github.com/aws/aws-application-networking-k8s/pkg/aws"
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/service/ram"
 	"os"
 
 	"github.com/aws/aws-sdk-go/service/vpclattice"
@@ -20,14 +23,30 @@ import (
 )
 
 const (
-	k8snamespace = "e2e-test"
+	k8snamespace                = "e2e-test"
+	secondaryAccountTestRoleArn = "SECONDARY_ACCOUNT_TEST_ROLE_ARN"
 )
 
+var k8sTestTags = map[string]*string{
+	anaws.TagBase + "TestSuite": aws.String("sharing"),
+}
+var k8sRamTestTags []*ram.Tag
 var testFramework *test.Framework
 var ctx context.Context
 var testGateway *gwv1.Gateway
 var testServiceNetwork *vpclattice.ServiceNetworkSummary
+var SecondaryAccountTestRoleArn string
+
 var _ = SynchronizedBeforeSuite(func() {
+	for key, value := range k8sTestTags {
+		k8sRamTestTags = append(k8sRamTestTags, &ram.Tag{
+			Key:   aws.String(key),
+			Value: value,
+		})
+	}
+
+	SecondaryAccountTestRoleArn = os.Getenv(secondaryAccountTestRoleArn)
+
 	vpcId := os.Getenv("CLUSTER_VPC_ID")
 	if vpcId == "" {
 		Fail("CLUSTER_VPC_ID environment variable must be set to run integration tests")
